Recovery email migration (#496)
* Refactoring migrations

* Fix last device activity

* Refactoring Recovery email

+++ electron_app/src/migrations/20180828163335_recoveryEmail.js
@@ -1,39 +1,19 @@
-const { Table, fieldTypes } = require('./../models');
-const { MEDIUM_STRING_SIZE } = fieldTypes;
+const { Table } = require('./../models');
 
-const createRecoveryEmailColumn = knex => {
-  return knex.schema.table(Table.ACCOUNT, table => {
-    table.string('recoveryEmail', MEDIUM_STRING_SIZE).defaultTo('');
-  });
-};
-
-const createRecoveryEmailConfirmedColumn = knex => {
-  return knex.schema.table(Table.ACCOUNT, table => {
-    table.boolean('recoveryEmailConfirmed').defaultTo(false);
-  });
-};
-
-exports.up = async (knex, Promise) => {
-  const checkAndCreateRecoveryEmailColumn = knex.schema
-    .hasColumn(Table.ACCOUNT, 'recoveryEmail')
-    .then(columnExists => {
-      return !columnExists && createRecoveryEmailColumn(knex);
-    });
-  const checkAndCreateIsVerifiedRecoveryEmailColumn = knex.schema
-    .hasColumn(Table.ACCOUNT, 'recoveryEmailConfirmed')
-    .then(columnExists => {
-      return !columnExists && createRecoveryEmailConfirmedColumn(knex);
-    });
-  return await Promise.all([
-    checkAndCreateRecoveryEmailColumn,
-    checkAndCreateIsVerifiedRecoveryEmailColumn
-  ]);
+exports.up = (knex, Promise) => {
+  return Promise.resolve(knex);
 };
 
 // On rollback
-exports.down = knex => {
-  return knex.schema.table(Table.ACCOUNT, table => {
-    table.dropColumn('recoveryEmail');
-    table.dropColumn('recoveryEmailConfirmed');
-  });
+exports.down = async knex => {
+  const hasRecoveryEmailColumn = await knex.schema.hasColumn(
+    Table.ACCOUNT,
+    'recoveryEmail'
+  );
+  if (hasRecoveryEmailColumn) {
+    await knex.schema.table(Table.ACCOUNT, table => {
+      table.dropColumn('recoveryEmail');
+      table.dropColumn('recoveryEmailConfirmed');
+    });
+  }
 };
